% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_GIS_cbpseg}
\alias{data_GIS_cbpseg}
\title{Chesapeake Bay Outline}
\format{
A simple feature collection with 6 features and 1 field.
\describe{
   \item{\code{combo}}{a numeric vector}
   \item{\code{geometry}}{a sfc_POLYGON}
}
}
\usage{
data_GIS_cbpseg
}
\description{
Official CBP GIS shapefile outline of Chesapeake Bay
}
\keyword{datasets}
