% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{run baytrends Shiny Example}
\usage{
runShiny()
}
\description{
Launches Shiny app for baytrends
}
\details{
The Shiny app based on the R package baytrends is included in the R 
package. This function launches that app.

The Shiny app is online at:
https://leppott.shinyapps.io/baytrendsmap
}
\examples{
\dontrun{
# Run Function
runShiny()
}
}
