% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_GIS_cbpseg}
\alias{data_GIS_cbpseg}
\title{Chesapeake Bay Outline}
\format{
A data frame with 274836 observations on the following 7 variables.
\describe{
  \item{long}{a numeric vector}
  \item{lat}{a numeric vector}
  \item{order}{a numeric vectorr}
  \item{hole}{a logical vector}
  \item{piece}{a factor with levels \code{1}}
  \item{id}{a character vector}
  \item{group}{a factor with levels}
}
}
\usage{
data_GIS_cbpseg
}
\description{
Official CBP GIS shapefile outline of Chesapeake Bay
}
\keyword{datasets}
